Added simplified implementation of factory method design pattern
