+++ src/main/java/design/object/creational/factorymethod/canonical/EncryptionAlgorithm.java
@@ -1,4 +1,4 @@
-package design.object.creational.factorymethod;
+package design.object.creational.factorymethod.canonical;
 
 /**
  * Product interface in terminology of factory method pattern
