+++ src/main/java/design/object/creational/factorymethod/canonical/IdeaEncryptor.java
@@ -1,12 +1,12 @@
-package design.object.creational.factorymethod;
+package design.object.creational.factorymethod.canonical;
 
 /**
  * Concrete creator which specifies what factory method does
  */
-public class TeaEncryptor extends Encryptor {
+public class IdeaEncryptor extends Encryptor {
 
     @Override
     public EncryptionAlgorithm getEncryptionAlgorithm() {
-        return new TeaEncryptionAlgorithm();
+        return new IdeaEncryptionAlgorithm();
     }
 }
