+++ src/main/java/design/object/creational/factorymethod/canonical/TeaEncryptor.java
@@ -1,12 +1,12 @@
-package design.object.creational.factorymethod;
+package design.object.creational.factorymethod.canonical;
 
 /**
  * Concrete creator which specifies what factory method does
  */
-public class IdeaEncryptor extends Encryptor {
+public class TeaEncryptor extends Encryptor {
 
     @Override
     public EncryptionAlgorithm getEncryptionAlgorithm() {
-        return new IdeaEncryptionAlgorithm();
+        return new TeaEncryptionAlgorithm();
     }
 }
