+++ src/main/java/design/object/creational/factorymethod/simplified/Multigon.java
@@ -1,0 +1,12 @@
+package design.object.creational.factorymethod.simplified;
+
+/**
+ * Concrete product in terms of factory method design pattern
+ */
+public class Multigon implements Polygon {
+
+    @Override
+    public void draw() {
+        // concrete drawing logic
+    }
+}
