+++ src/main/java/design/object/creational/factorymethod/simplified/Polygon.java
@@ -1,0 +1,12 @@
+package design.object.creational.factorymethod.simplified;
+
+/**
+ * Product interface in terms of factory method design pattern
+ */
+public interface Polygon {
+
+    /**
+     * Draws itself on a screen
+     */
+    void draw();
+}
