+++ src/main/java/design/object/creational/factorymethod/simplified/Square.java
@@ -1,0 +1,12 @@
+package design.object.creational.factorymethod.simplified;
+
+/**
+ * Concrete product in terms of factory method design pattern
+ */
+public class Square implements Polygon {
+
+    @Override
+    public void draw() {
+        // concrete drawing logic
+    }
+}
